Check and fix printf style arguments to die() and pdie()
Add the "format" gcc style attribute to the declarations of die() and
pdie() so that the compiler knows how to check their arguments for
correctness.

Fix all problematic calls detected by this.

+++ src/boot.c
@@ -416,7 +416,8 @@ void read_boot(DOS_FS * fs)
 	 * much clusers otherwise. */
 	fs->fat_bits = (fs->data_clusters >= FAT12_THRESHOLD) ? 16 : 12;
 	if (fs->data_clusters >= FAT16_THRESHOLD)
-	    die("Too many clusters (%lu) for FAT16 filesystem.", fs->data_clusters);
+	    die("Too many clusters (%lu) for FAT16 filesystem.",
+		    (unsigned long)fs->data_clusters);
 	check_fat_state_bit(fs, &b);
     } else {
 	/* On Atari, things are more difficult: GEMDOS always uses 12bit FATs
